requirements.txt
boto3==1.9.239
botocore==1.12.239
capstone==4.0.1
command==0.1.0
editdistance==0.6.0
fairscale==0.4.8
fastBPE==0.1.0
fvcore==0.1.5.post20220512
gensim==3.8.1
gossip==2.4.0
hydra-core==1.2.0
nltk==3.4.5
omegaconf
psutil==5.8.0
pyelftools==0.29
PyYAML==6.0
regex==2021.7.6
requests
sacrebleu==2.0.0
sacremoses==0.0.46
sentencepiece==0.1.95
setuptools==57.5.0
SoundFile==0.10.3.post1
spacy==3.3.1
subword_nmt==0.3.8
tensorboardX==2.5.1
tokenizers==0.10.3
tqdm==4.62.3
transformers==4.16.2
tweepy==4.10.1
pyhidra==0.4.1
construct==2.9.52
pefile==2023.2.7
